PDJB-1279: Extract shared base page for org company number page objects
